+++ s3pypi/storage.py
@@ -1,48 +1,45 @@
+from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Optional
+from typing import Dict, Optional
 
 import boto3
 import botocore
-from botocore.config import Config
+from botocore.config import Config as BotoConfig
+from mypy_boto3_s3.service_resource import Object
 
 from s3pypi.index import Index
+
+
+@dataclass
+class S3Config:
+    bucket: str
+    prefix: Optional[str] = None
+    endpoint_url: Optional[str] = None
+    put_kwargs: Dict[str, str] = field(default_factory=dict)
+    unsafe_s3_website: bool = False
+    no_sign_request: bool = False
 
 
 class S3Storage:
     root = "/"
     _index = "index.html"
 
-    def __init__(
-        self,
-        session: boto3.session.Session,
-        bucket: str,
-        prefix: Optional[str] = None,
-        acl: Optional[str] = None,
-        s3_endpoint_url: Optional[str] = None,
-        s3_put_args: Optional[dict] = None,
-        unsafe_s3_website: bool = False,
-        no_sign_request: bool = False,
-    ):
+    def __init__(self, session: boto3.session.Session, cfg: S3Config):
         _config = None
-        if no_sign_request:
-            _config = Config(signature_version=botocore.session.UNSIGNED)
+        if cfg.no_sign_request:
+            _config = BotoConfig(signature_version=botocore.session.UNSIGNED)  # type: ignore
 
-        self.s3 = session.resource("s3", endpoint_url=s3_endpoint_url, config=_config)
-        self.bucket = bucket
-        self.prefix = prefix
-        self.index_name = self._index if unsafe_s3_website else ""
-        self.put_kwargs = dict(
-            ACL=acl or "private",
-            **(s3_put_args or {}),
-        )
+        self.s3 = session.resource("s3", endpoint_url=cfg.endpoint_url, config=_config)
+        self.index_name = self._index if cfg.unsafe_s3_website else ""
+        self.cfg = cfg
 
-    def _object(self, directory: str, filename: str):
+    def _object(self, directory: str, filename: str) -> Object:
         parts = [directory, filename]
         if parts == [self.root, self.index_name]:
             parts = [self._index]
-        if self.prefix:
-            parts.insert(0, self.prefix)
-        return self.s3.Object(self.bucket, key="/".join(parts))
+        if self.cfg.prefix:
+            parts.insert(0, self.cfg.prefix)
+        return self.s3.Object(self.cfg.bucket, key="/".join(parts))
 
     def get_index(self, directory: str) -> Index:
         try:
@@ -54,26 +51,26 @@ class S3Storage:
     def build_root_index(self) -> Index:
         paginator = self.s3.meta.client.get_paginator("list_objects_v2")
         result = paginator.paginate(
-            Bucket=self.bucket,
-            Prefix=self.prefix or "",
+            Bucket=self.cfg.bucket,
+            Prefix=self.cfg.prefix or "",
             Delimiter="/",
         )
-        n = len(self.prefix) + 1 if self.prefix else 0
+        n = len(self.cfg.prefix) + 1 if self.cfg.prefix else 0
         dirs = (p.get("Prefix")[n:] for p in result.search("CommonPrefixes"))
         return Index(dict.fromkeys(dirs))
 
-    def put_index(self, directory: str, index: Index):
+    def put_index(self, directory: str, index: Index) -> None:
         self._object(directory, self.index_name).put(
             Body=index.to_html(),
             ContentType="text/html",
             CacheControl="public, must-revalidate, proxy-revalidate, max-age=0",
-            **self.put_kwargs,
+            **self.cfg.put_kwargs,  # type: ignore
         )
 
-    def put_distribution(self, directory: str, local_path: Path):
+    def put_distribution(self, directory: str, local_path: Path) -> None:
         with open(local_path, mode="rb") as f:
             self._object(directory, local_path.name).put(
                 Body=f,
                 ContentType="application/x-gzip",
-                **self.put_kwargs,
+                **self.cfg.put_kwargs,  # type: ignore
             )