+++ tests/integration/test_main.py
@@ -44,15 +44,19 @@ def test_main_upload_package_exists(chdir, data_dir, s3_bucket, caplog):
     dist = "dists/foo-0.1.0.tar.gz"
 
     with chdir(data_dir):
-        for _ in range(2):
-            s3pypi(dist, "--bucket", s3_bucket.name)
+        s3pypi(dist, "--bucket", s3_bucket.name)
+        s3pypi(dist, "--bucket", s3_bucket.name)
+
+        with pytest.raises(SystemExit, match="ERROR: Found 1 existing files on S3"):
+            s3pypi(dist, "--strict", "--bucket", s3_bucket.name)
+
         s3pypi(dist, "--force", "--bucket", s3_bucket.name)
 
     msg = "foo-0.1.0.tar.gz already exists! (use --force to overwrite)"
     success = (__prog__, logging.INFO, "Uploading " + dist)
     warning = (__prog__, logging.WARNING, msg)
 
-    assert caplog.record_tuples == [success, warning, success]
+    assert caplog.record_tuples == [success, warning, warning, success]
 
 
 @pytest.mark.parametrize(