+++ tests/integration/test_storage.py
@@ -1,12 +1,14 @@
 from s3pypi.index import Index
-from s3pypi.storage import S3Storage
+from s3pypi.storage import S3Config, S3Storage
 
 
 def test_index_storage_roundtrip(boto3_session, s3_bucket):
     directory = "foo"
     index = Index({"bar": None})
 
-    storage = S3Storage(boto3_session, s3_bucket.name)
+    cfg = S3Config(bucket=s3_bucket.name)
+    storage = S3Storage(boto3_session, cfg)
+
     storage.put_index(directory, index)
     got = storage.get_index(directory)
 
@@ -14,9 +16,9 @@ def test_index_storage_roundtrip(boto3_session, s3_bucket):
 
 
 def test_prefix_in_s3_key(boto3_session):
-    prefix = "1234567890"
+    cfg = S3Config(bucket="example", prefix="1234567890")
+    storage = S3Storage(boto3_session, cfg)
 
-    storage = S3Storage(boto3_session, bucket="example", prefix=prefix)
     obj = storage._object(directory="foo", filename="bar")
 
-    assert obj.key.startswith(prefix + "/")
+    assert obj.key.startswith(cfg.prefix + "/")